requirements_dev.txt
-r requirements.txt
pep8>=1.7.1
flake8>=3.6.0
black>=18.9b0
pylint>=2.2.2
pytest>=4.0.1
pytest-django>=3.4.4
pytest-mock>=1.10
pytest-sugar>=0.9.1
pytest-cov>=2.5.1
pytest-xdist>=1.24.1
codecov==2.0.15
factory_boy>=2.11.1
pdbpp>=0.9.2
hypothesis>=3.82.5
wdb
